moved structure util to its own file (to avoid running cleantests twice by importing it)
