+++ .gitignore
@@ -3,7 +3,7 @@ __pycache__/
 .conda/
 .idea/
 build/
-test_clean_scratchspace/
+*_scratchspace/
 workspace/
 default_*_benchbase_config_*.xml
 *.egg-info/
